+++ packages/docs-infra/src/pipeline/hastUtils/index.ts
@@ -1,3 +1,4 @@
 export * from './hastCompression';
 export * from './hastUtils';
+export { getHastTextContent, getShallowTextContent } from './getHastTextContent';
 export { stripHighlightingSpans } from './stripHighlightingSpans';
