+++ packages/docs-infra/src/pipeline/loadServerTypes/hastTypeUtils.ts
@@ -6,20 +6,9 @@
  */
 
 import type { Root as HastRoot, Element, Text, RootContent } from 'hast';
-import { compressHast } from '../hastUtils';
+import { compressHast, getHastTextContent, getShallowTextContent } from '../hastUtils';
 
-/**
- * Extracts all text content from a HAST node recursively.
- */
-export function getHastTextContent(node: HastRoot | Element | Text | RootContent): string {
-  if (node.type === 'text') {
-    return node.value || '';
-  }
-  if ('children' in node && Array.isArray(node.children)) {
-    return node.children.map((child) => getHastTextContent(child as RootContent)).join('');
-  }
-  return '';
-}
+export { getHastTextContent, getShallowTextContent };
 
 /**
  * Checks if a HAST element has a specific CSS class.
@@ -34,19 +23,6 @@ export function hasClass(element: Element, className: string): boolean {
     return classes.split(' ').includes(className);
   }
   return false;
-}
-
-/**
- * Gets the direct text content of a HAST element (non-recursive, first level only).
- */
-export function getShallowTextContent(element: Element): string {
-  let text = '';
-  for (const child of element.children) {
-    if (child.type === 'text') {
-      text += child.value;
-    }
-  }
-  return text;
 }
 
 /**
